requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
requests==2.26.0
python-dotenv==0.19.0
huggingface-hub==0.19.4
transformers==4.36.2
replicate==0.22.0